refactor(VtkTwoView): Switch to vue 3 impl
